refactor : 검증하는 모듈을 분리
